Continued work on the Interpreter
HEUREKA!
This commit contains the first working version of the RaychelScript
interpreter. There is still a lot to do but it feels good to see your
scripting language go from a file to a computed value :)
This interpreter walks the AST and populates special state variables as it
goes. There are a few quirks with my approach (e.g. the state variable
_load_references which decides whether variable references should load
the value or the descriptor).
And for my dear friend @tantarusss: Bogos Binted