Added parsing support for conditional constructs
The Lexer will now produce conditional_header and condition_footer tokens. When the parser encounters a construct that matches CONDITIONAL_HEADER EXPRESSION it will push a new scope onto the newly introduced scope stack. If there is a scope present on the stack, the parser will push all parsed nodes into that scope.
If it encounters a construct like CONDITIONAL_FOOTER, the current scope will be pushed onto the stack