Added PipeResult class
This class greatly improves the Pipes API by unifying error handling for all types.
A PipeResult<T> holds either an object of type T (the expected result) or any error that occured within the pipeline.
If for example the lexing step fails, the error returned will be the lexers error instead of somebodys no_input.
We also provide the convenience function log_if_error that will log the error that occured, if any