VM: updated execution loop
According to callgrind, we were spending a lot of time comparing iterators in the main loop condition. We don't need to do this since the last instruction will always set the HALT flag. To ensure that the last instruction is actually HLT and we don't run off the end of our vector, we added a new check and error code