VM: implement opcode handling using the computed gotos extension
This *NONSTANDARD* compiler extension allows storing the address of a label and jumping to a dynamic label. This can improve the performance of bytecode interpreters.

This option is disabled by default and can be enabled by defining the macro RAYCHELSCRIPT_VM_USE_COMPUTED_GOTO to 1

Note that enabling this option disables dynamic floating-point exception handling